requirements.txt
Pillow==9.5.0
l10n~=0.1.0
reedsolo==2.0.13
